requirements.txt
Flask==1.1.4
Jinja2==2.11.3
markupsafe==2.0.1
gunicorn
googlesearch-python==1.1.0
requests==2.25.1
flask-googlemaps==0.4.1
newsapi-python==0.2.6